place数据保存

+++ app/src/main/java/com/kunkun/sunnyweather/logic/model/DailyResponse.kt
@@ -86,7 +86,7 @@ data class DailyResponse(
         val avg: Double,
         val date: String,
         val max: Double,
-        val min: Int,
+        val min: Double,
         val probability: Int
     )
